Added Basic authentication for anything other than a get request

+++ swagger_server/controllers/pages_controller.py
@@ -1,6 +1,7 @@
 import connexion
 import six
 import os
+from .auth import authenticate
 
 from flask import jsonify
 from flask import send_file
@@ -21,9 +22,17 @@ def add_page(file, filename):  # noqa: E501
 
     :rtype: None
     """
+    if not authenticate(connexion.request):
+        return jsonify({'status': 'user/pass authorization is missing or invalid'}), 401
+        
     try:
         if not os.path.exists(static_directory):
             os.makedirs(static_directory)
+
+        local_file = Path(f'{static_directory}/{filename}')
+        if local_file.exists():
+            raise ValueError('file already exists')
+
         file.save(f'{static_directory}/{filename}')
         file.close()
         return jsonify({'status': 'File Added'}), 201
@@ -41,6 +50,9 @@ def delete_page(filename):  # noqa: E501
 
     :rtype: None
     """
+    if not authenticate(connexion.request):
+        return jsonify({'status': 'user/pass authorization is missing or invalid'}), 401
+
     try:
         os.remove(f'{static_directory}/{filename}')
         return jsonify({'status': 'OK'}), 200
@@ -76,6 +88,9 @@ def update_page(file, filename):  # noqa: E501
 
     :rtype: None
     """
+    if not authenticate(connexion.request):
+        return jsonify({'status': 'user/pass authorization is missing or invalid'}), 401
+        
     try:
         os.remove(f'{static_directory}/{filename}')
         file.save(f'{static_directory}/{filename}')
